first changes to support a native build under Windows in Visual Studio 2008 simply define WINDOWS in your compileroptions
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@332 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/basis/linearboundary/algorithm_sweep/LaplaceDownLinearBoundary.hpp
@@ -137,8 +137,7 @@ protected:
 			GridStorage::index_type::index_type i;
 
 			index.get(dim, l, i);
-
-			double h = 1/pow(2.0, l);
+			double h = 1/pow(2.0, static_cast<int>(l));
 
 			// integration
 			result[seq] = (  h * (fl+fr)/2.0
