+++ src/sgpp/basis/linearboundaryUScaled/algorithm_sweep/LaplaceDownLinearBoundaryUScaled.hpp
@@ -139,7 +139,7 @@ protected:
 
 		index.get(dim, l, i);
 
-		double h = 1/pow(2.0, l);
+		double h = 1/pow(2.0, static_cast<int>(l));
 
 		// integration
 		result[seq] = (  h * (fl+fr)/2.0
