+++ src/sgpp/basis/modlinear/algorithm_sweep/LaplaceDownGradientModLinear.hpp
@@ -93,7 +93,7 @@ protected:
 		index.get(dim, l, i);
 
 		double alpha_value = source[seq];
-		double ht = pow(2.0, l);
+		double ht = pow(2.0, static_cast<int>(l));
 		double f_local = 0.0;
 
 		// level 1, constant function
